Refactored almost all of score period modals. This also marks the start of the minimal comments period
